refactor to improve write file
