+++ src/store/electron/selectors/desktopState.ts
@@ -1,0 +1,7 @@
+import { ElectronState } from '@/store/electron/initialState';
+
+const usePath = (s: ElectronState) => s.appState.userPath;
+
+export const desktopStateSelectors = {
+  usePath,
+};
