+++ src/store/electron/selectors/index.ts
@@ -1,1 +1,2 @@
+export * from './desktopState';
 export * from './sync';
